refactor(workspace): extract instructions and skills crates
