+++ Cargo.toml
@@ -1,3 +1,20 @@
+[workspace]
+members = [".", "crates/config", "crates/instructions", "crates/skills"]
+resolver = "2"
+
+[workspace.package]
+version = "0.1.0"
+edition = "2021"
+license = "Apache-2.0"
+
+[workspace.dependencies]
+anyhow = "1.0"
+dirs = "6.0"
+secrecy = { version = "0.10", features = ["serde"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+tempfile = "3.17"
+
 [package]
 name = "rara"
 version = "0.1.0"
@@ -21,6 +38,9 @@ rand = "0.8"
 base64 = "0.22"
 sha2 = "0.10"
 secrecy = { version = "0.10", features = ["serde"] }
+rara-config = { path = "crates/config" }
+rara-instructions = { path = "crates/instructions" }
+rara-skills = { path = "crates/skills" }
 urlencoding = "2.1"
 url = "2.5"
 ratatui = "0.30"
