+++ crates/config/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "rara-config"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+secrecy.workspace = true
+serde.workspace = true
+serde_json.workspace = true
