+++ crates/instructions/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "rara-instructions"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+rara-config = { path = "../config" }
+
+[dev-dependencies]
+tempfile.workspace = true
