+++ crates/instructions/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod prompt;
+pub mod workspace;
+
+pub use prompt::*;
+pub use workspace::*;
